Hough transform
This is a lane detection algorithm developed by Timur and me using Hough Transform. Timur added the functionality of tracking the distance between the points and we came up with the algorithm to keep the rover straight along with the Arduino code for the encoders to keep the motor straight. The code will be uploaded by Quang after he is made some necessary comments and formatting.